requirements.txt
aiobotocore==2.2.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
botocore==1.24.21
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cmdstanpy==1.0.4
convertdate==2.4.0
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.28
Deprecated==1.2.13
dill==0.3.4
ephem==4.1.3
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
hijri-converter==2.2.3
holidays==0.13
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.8.0
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.6.0
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
lightgbm==3.3.2
LunarCalendar==0.0.9
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.1
memory-profiler==0.60.0
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.7.0
pluggy==1.0.0
prophet==1.1
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
py4j==0.10.9.5
pycodestyle==2.8.0
pycparser==2.21
PyMeeus==0.5.11
pyparsing==3.0.8
pyrsistent==0.18.1
pystan==2.19.1.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rq==1.10.1
s3fs==2022.3.0
salesforce-merlion==1.2.5
scikit-learn==1.0.2
scipy==1.8.0
setuptools-git==1.2
simple-salesforce==1.11.6
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
# Editable install with no version control (src-configs==1.0)
-e .
statsmodels==0.13.2
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.2.0
ujson==5.4.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.0
yarl==1.7.2
zeep==4.1.0
zipp==3.8.0